Tighten integer types in public API
Per the audit naming scheme: bool returns -> uint32_t, indices ->
uint32_t (where negative is meaningless), int versions of major /
minor / enable / transparent -> int32_t. No new diagnostics under
-Wall -Wextra.

Sites touched:
- mkfw_gl_major / minor / transparent file-statics: int -> int32_t
- mkfw_set_gl_version / mkfw_set_transparent parameters
- mkfw_parse_version (return + parameters)
- mkfw_query_max_gl_version (return + parameters) on both backends
- mkfw_window_is_key_pressed / was_key_released / is_button_pressed
  / was_button_released returns: int / uint8_t -> uint32_t
- mkfw_window_should_close / is_minimized / is_maximized returns:
  int32_t -> uint32_t (on both backends)
- All joystick query helpers: int -> uint32_t for both indices and
  returns. The < 0 bounds checks are removed (tautological for
  unsigned and tripped -Wtype-limits)
- mkfw_joystick_rumble_platform pad_index: int -> uint32_t
- mkfw_futex_wait / wake: int -> int32_t (timer internal)

Last commit of the mechanical rename pass (step 3).

+++ mkfw_win32_joystick.c
@@ -128,10 +128,10 @@ static void mkfw_joystick_update(void) {
 }
 
 // [=]===^=[ mkfw_joystick_rumble_platform ]======================================================[=]
-static void mkfw_joystick_rumble_platform(int pad_index, float low_freq, float high_freq, uint32_t duration_ms) {
+static void mkfw_joystick_rumble_platform(uint32_t pad_index, float low_freq, float high_freq, uint32_t duration_ms) {
 	(void)duration_ms;
 	XINPUT_VIBRATION vibration;
 	vibration.wLeftMotorSpeed = (WORD)(low_freq * 65535.0f);
 	vibration.wRightMotorSpeed = (WORD)(high_freq * 65535.0f);
-	XInputSetState((DWORD)pad_index, &vibration);
+	XInputSetState(pad_index, &vibration);
 }
